+++ src/app/exams/planner/ExamDataEditor/update.ts
@@ -4,12 +4,13 @@ import { Auth } from "@/lib/auth";
 import { schema } from "./schema";
 import { prisma } from "@/lib/db";
 import { revalidatePath } from "next/cache";
-import { getThaiDate } from "@/utils/date";
+import { timeStringToMinutes } from "@/utils/date";
 import UserExamModel from "@/mongoose/model/UserExam";
 import LogModel, { LogAction } from "@/mongoose/model/Log";
 import { ExamSchedule } from "@/lib/generated/prisma";
 import { TableSource } from "@/mongoose/enum/TableSource";
 import UserExamPlannerModel from "@/mongoose/model/UserExamPlanner";
+import { envServer } from "@/env/server.mjs";
 type ResponseCode =
   | "UNAUTHORIZED"
   | "NOT_FOUND"
@@ -23,24 +24,53 @@ interface ResponseData {
 }
 
 export async function update(payload: string): Promise<ResponseData> {
-  try {
-    const isAuth = await Auth();
-    if (!isAuth) {
-      return { message: "โปรดเข้าสู่ระบบ", code: "UNAUTHORIZED" };
-    }
-    const valid = schema.safeParse(JSON.parse(payload));
-    if (!valid.success) {
-      console.log(valid.error.issues[0]);
-      return { message: valid.error.issues[0].message, code: "INVALID_FORMAT" };
-    }
+  const isAuth = await Auth();
+  if (!isAuth) {
+    return { message: "โปรดเข้าสู่ระบบ", code: "UNAUTHORIZED" };
+  }
+  const valid = schema.safeParse(JSON.parse(payload));
+  if (!valid.success) {
+    return { message: valid.error.issues[0].message, code: "INVALID_FORMAT" };
+  }
 
-    const { id, subjectCode, room, date, timeFrom, timeTo, sectionCode } =
-      valid.data;
-    const { stdCode } = isAuth;
-    return await prisma.$transaction(async (tx) => {
-      const isRegistered = await tx.registeredCourse.findUnique({
-        where: { id, deletedAt: null },
-        include: { CourseSchedule: true },
+  const { id, subjectCode, room, date, sectionCode } = valid.data;
+  const { id: stdCode } = isAuth.session;
+  const [timeFrom, timeTo] = [
+    timeStringToMinutes(valid.data.timeFrom),
+    timeStringToMinutes(valid.data.timeTo),
+  ];
+  return await prisma.$transaction(async (tx) => {
+    const isRegistered = await tx.registeredCourse.findUnique({
+      where: { id, deletedAt: null },
+      include: { CourseSchedule: true },
+    });
+    if (!isRegistered || !isRegistered.CourseSchedule) {
+      return { message: "ไม่พบว่าคุณมีรายวิชานี้ในตาราง", code: "NOT_FOUND" };
+    }
+    const isExistExam = await tx.examSchedule.findFirst({
+      where: {
+        stdCode,
+        subjectCode,
+        deletedAt: null,
+        reportBy: "STUDENT",
+        sectionType: isRegistered.CourseSchedule.sectionType,
+      },
+    });
+    let updated: {
+      type: keyof typeof LogAction;
+      data: ExamSchedule;
+    } | null = null;
+    if (isExistExam) {
+      const r = await tx.examSchedule.update({
+        where: {
+          id: isExistExam.id,
+        },
+        data: {
+          room,
+          date,
+          timeFrom,
+          timeTo,
+        },
       });
       if (!isRegistered || !isRegistered.CourseSchedule) {
         return { message: "ไม่พบข้อมูลรายวิชานี้", code: "NOT_FOUND" };
@@ -49,8 +79,13 @@ export async function update(payload: string): Promise<ResponseData> {
         where: {
           stdCode,
           subjectCode,
-          deletedAt: null,
+          sectionCode,
+          room,
+          date,
+          timeFrom,
+          timeTo,
           reportBy: "STUDENT",
+          sectionType: isRegistered.CourseSchedule.sectionType,
         },
       });
       const formattedTime = `${timeFrom.replace(":", ".")}-${timeTo.replace(
@@ -95,35 +130,35 @@ export async function update(payload: string): Promise<ResponseData> {
         updated = { type: "CREATE_EXAM", data: insertExamSchedule };
       }
 
-      // Clear MongoDB cache
-      await Promise.all([
-        UserExamModel.findOneAndDelete({ stdCode }),
-        UserExamPlannerModel.findOneAndDelete({ stdCode }),
-      ]);
+    if (envServer.NODE_ENV === "production") {
+      // Update MongoDB cache
+      await UserExamModel.findOneAndDelete({ stdCode });
+      await UserExamPlannerModel.findOneAndDelete({ stdCode });
+    }
 
-      if (updated) {
-        await LogModel.create({
-          stdCode,
-          action: updated.type,
-          tableSource: TableSource.EXAM_SCHEDULE,
-          pkid: updated.data.id,
-          dataJson: JSON.stringify({
-            ...updated.data,
-            id: undefined,
-            dateTh: undefined,
-            createdAt: undefined,
-            deletedAt: undefined,
-            studentIdRange: undefined,
-            isCorrectCount: undefined,
-            status: undefined,
-            stdCode: undefined,
-            reportBy: undefined,
-            sectionType: undefined,
-            sectionCode: undefined,
-            subjectCode: undefined,
-          }),
-        });
-      }
+    if (updated) {
+      await LogModel.create({
+        stdCode,
+        action: updated.type,
+        tableSource: TableSource.EXAM_SCHEDULE,
+        pkid: updated.data.id,
+        dataJson: JSON.stringify({
+          ...updated.data,
+          id: undefined,
+          createdAt: undefined,
+          deletedAt: undefined,
+          studentIdRange: undefined,
+          isCorrectCount: undefined,
+          isIncorrectCount: undefined,
+          status: undefined,
+          stdCode: undefined,
+          reportBy: undefined,
+          sectionType: undefined,
+          sectionCode: undefined,
+          subjectCode: undefined,
+        }),
+      });
+    }
 
       revalidatePath("/exams/planner");
       revalidatePath("/exams");
